Split VRM files
